pyproject.toml
[tool.poetry]
name = "DiscoTube"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^2.0.1"
express = "^1.0"
conwech = "^2020.8"
colorama = "^0.4.4"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
